Fix edge case where color temp is reset to TEMPERATURE_ZERO in delta mode
This quirk is explained in the README. Basically if the brightness that's passed to
get_sct_for_screen() is less than or equal to zero, the temperature is 'set' to 0.
This means that if we were using delta mode to decrease the brightness, the temperature
information will be lost whenever the brightness becomes less than or equal to 0.

Before this commit, this would manifest itself by forcing the temperature to 700K.
For example, doing `xsct 6500 -1; xsct -d 0 1` would result in a red screen, even though
we didn't intend to change the color temperature.

Ideally, the temperature information should stay the same if we are only shifting the brightness
and we reach 0.0, but I couldn't think of a straightforward way to implement this.

Also add a fix to make sure temp and brightness are included when using the delta flag