Address edge case where temp info is reset when brightness goes below 0.0 in delta mode (#45)
Before this commit the temperature would be reset to `TEMPERATURE_ZERO`
if the brightness was reduced to 0.0 in delta mode. Now it resets to
`TEMPERATURE_NORM`. I noticed this because I have a keyboard shortcut
which can quickly reduce and increase brightness using the delta mode.

If you want to test it, you can do: `xsct 6500 -1; xsct -d 0 1`. This
results in a red screen, even though it wasn't intended to change the
temperature from 6500.

Now the temperature is reset to 6500, which isn't the perfect solution,
but at least it's somewhat aligned to passing 0 as a parameter to `xsct`
in absolute mode.

Also add small fix to make sure both temp and brightness are provided
when in delta mode.

---------

Co-authored-by: Fabian Foerg <3429782+faf0@users.noreply.github.com>